requirements.txt
google-genai
python-dotenv
slack-bolt